final/backend/requirements.txt
Flask
Flask-Cors
psycopg2-binary
SQLAlchemy
requests
numpy
pandas
prophet
redis
weaviate-client
python-dotenv
